+++ src/app/models/reservation.ts
@@ -1,16 +1,20 @@
+
 import { Boat } from './boat';
 import { User } from './user';
 export class Reservation {
-    ReservationId: string;
-    CreatedBy: string;
-    User: User;
-    ReservedBoat: string;
-    Boat: Boat;
-    StartDateTime: Date;
-    EndDateTime: Date;
-    MemberCrew: User[];
-    NonMemberCrew: User[];
-    Itinerary: string;
-    AllocatedCredit: number;
-    AllocatedHours: number;
+    constructor(
+        ReservationId?: string,
+        CreatedBy?: string,
+        User?: User,
+        ReservedBoat?: string,
+        Boat?: Boat,
+        StartDateTime?: Date,
+        EndDateTime?: Date,
+        MemberCrew?: User[],
+        NonMemberCrew?: User[],
+        Itinerary?: string,
+        AllocatedCredit?: number,
+        AllocatedHours?: number
+    ){}
+
 }